fix: Minor code changes to delete_trip and delete_route views

+++ bookings/views.py
@@ -182,20 +182,29 @@ def edit_trip(request, trip_id):
 
 # Superuser can delete a Route
 def delete_route(request, route_id):
-    route = Route.objects.get(id=route_id)
 
-    # Check that user really wants to delete this route
-    if request.method == 'POST':
-        route.delete()
+    # This page can only be accessed by a superuser
+    if request.user.is_superuser:
+        route = Route.objects.get(id=route_id)
+
+        # Check that user really wants to delete this route
+        if request.method == 'POST':
+            route.delete()
+            messages.success(request, (
+                'Success! The route has been deleted from the database'))
+            return redirect('admin-panel')
+
+        context = {
+                'route': route
+                }
+
+        return render(request, 'delete_route.html', context)
+
+# For non-superusers trying to access the page
+    else:
         messages.success(request, (
-            'Success! The route has been deleted from the database'))
-        return redirect('admin-panel')
-
-    context = {
-            'route': route
-            }
-
-    return render(request, 'delete_route.html', context)
+            'Access denied. Please sign in as an admin.'))
+        return redirect('home')
 
 
 # Superuser can delete a Trip
